Remove deprecated Lambda runtimes python3_7, java8, go1_x, and provided
