+++ tests/aws/services/lambda_/functions/common/echo/provided_go/src/go.mod
