+++ tests/aws/services/lambda_/functions/common/echo/provided_go/src/main.go
