+++ tests/aws/services/lambda_/functions/common/endpointinjection/java8.al2/Makefile
