+++ tests/aws/services/lambda_/functions/common/endpointinjection/java8.al2/src/build.gradle
@@ -3,8 +3,9 @@ plugins {
 }
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
+    // Target oldest tested Java runtime
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
 }
 
 repositories {
@@ -12,8 +13,14 @@ repositories {
 }
 
 dependencies {
+    // AWS SDK v1: https://docs.aws.amazon.com/sdk-for-java/v1/developer-guide/welcome.html
+    implementation platform('com.amazonaws:aws-java-sdk-bom:1.12.599')
+    // Lambda and SQS
+    implementation 'com.amazonaws:aws-java-sdk-lambda'
+    implementation 'com.amazonaws:aws-java-sdk-sqs'
+
     // AWS SDK v2: https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/home.html
-    implementation platform('software.amazon.awssdk:bom:2.21.33')
+    implementation platform('software.amazon.awssdk:bom:2.21.35')
     // SQS and HTTP clients
     implementation 'software.amazon.awssdk:sqs'
     compileOnly 'software.amazon.awssdk:apache-client'
