+++ tests/aws/services/lambda_/functions/common/endpointinjection/provided_go/src/main.go
