+++ tests/aws/services/lambda_/functions/common/introspection/provided_go/Makefile
