+++ tests/aws/services/lambda_/functions/common/introspection/provided_go/src/go.mod
