+++ tests/aws/services/lambda_/functions/common/uncaughtexception/provided_go/src/go.mod
