+++ tests/aws/services/lambda_/functions/common/uncaughtexception/provided_go/src/main.go
